Disable full daily run
This runs the full run only monthly, and removes the check for whether
it is the first day of the month. This means we can more easily trigger
a full run any time as needed.

The full run was previously done daily only to generate a list of all
ingested items and all items that Zephir had metadata for, but this list
is only used monthly by ingest:
https://github.com/hathitrust/feed_internal/blob/main/bin/jobs/feed.monthly/zephir_diff.pl